# Resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API Development Document with current endpoints hosted on l2lc-dev",
    "version": "1.0.1",
    "title": "l2lc API"
  },
  "host": "138.197.88.198:8080",
  "basePath": "/l2lc/api",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/employee/create": {
      "post": {
        "tags": [
          "Employee"
        ],
        "summary": "Add a new employee",
        "description": "",
        "operationId": "addEmployee",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Employee object to be added to database",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Create_Employee_Input"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/Create_Employee_Output"
            }
          },
          "406": {
            "description": "Not Acceptable input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/employee/update/{id}": {
      "post": {
        "tags": [
          "Employee"
        ],
        "summary": "Add a new employee",
        "description": "",
        "operationId": "updateEmployeeById",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "required": true,
            "description": "Numeric Id of the employee"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Employee object to be added to database",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Create_Employee_Input"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK, update successful",
            "schema": {
              "$ref": "#/definitions/Create_Employee_Output"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "406": {
            "description": "Not Acceptable input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/employee/{id}": {
      "get": {
        "tags": [
          "Employee"
        ],
        "summary": "Get an employee's information from their id",
        "operationId": "getEmployeeById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "required": true,
            "description": "Numeric Id of the employee"
          }
        ],
        "responses": {
          "200": {
            "description": "Employee was successfully retrieved",
            "schema": {
              "$ref": "#/definitions/Create_Employee_Output"
            }
          },
          "404": {
            "description": "Employee Not found"
          }
        }
      }
    },
    "/employee/remove/{id}": {
      "delete": {
        "tags": [
          "Employee"
        ],
        "summary": "Remove an employee",
        "operationId": "removeEmployeeById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "required": true,
            "description": "Numeric Id of the employee"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Employee Not Found"
          }
        }
      }
    },
    "/employee": {
      "get": {
        "tags": [
          "Employee"
        ],
        "summary": "Get All Employees",
        "operationId": "getAllEmployees",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "type": "integer",
            "required": false,
            "description": "The amount of maximum amount of items to be returned"
          },
          {
            "in": "query",
            "name": "offset",
            "type": "integer",
            "required": false,
            "description": "The offset to be used when obtaining items."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Create_Employee_Output"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Create_Employee_Input": {
      "type": "object",
      "required": [
        "firstName",
        "lastName"
      ],
      "properties": {
        "firstName": {
          "type": "string",
          "example": "Joel"
        },
        "lastName": {
          "type": "string",
          "example": "Bornemann"
        }
      }
    },
    "Create_Employee_Output": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 12
        },
        "firstName": {
          "type": "string",
          "example": "Joel"
        },
        "lastName": {
          "type": "string",
          "example": "Bornemann"
        }
      }
    },
...
    "Error": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "NOT_ACCEPTABLE"
        },
        "errorMessages": {
          "type": "array",
          "items": {
            "type": "string",
            "example": "Some Error Message"
          }
        }
      }
    }
  }
}